fix after hyperspace comments
