+++ packages/fiori-generator-shared/package.json
@@ -29,6 +29,7 @@
         "!dist/**/*.map"
     ],
     "dependencies": {
+        "@sap-ux/axios-extension": "workspace:*",
         "@sap-ux/btp-utils": "workspace:*",
         "@sap-ux/project-access": "workspace:*",
         "@sap-ux/telemetry": "workspace:*",
@@ -48,7 +49,6 @@
         "@types/vscode": "1.110.0",
         "@types/yeoman-environment": "2.10.11",
         "@types/yeoman-generator": "5.2.14",
-        "@sap-ux/axios-extension": "workspace:*",
         "@sap-ux/logger": "workspace:*",
         "@sap-ux/store": "workspace:*"
     },
